+++ smos-web-server/src/Smos/Web/Server/TUI.hs
@@ -12,7 +12,11 @@ import Conduit
 import Data.Aeson as JSON
 import Data.ByteString (ByteString)
 import qualified Data.ByteString.Lazy as LB
+import qualified Data.Conduit.Combinators as C
+import qualified Data.Text as T
 import Data.Text (Text)
+import qualified Data.Text.Encoding as TE
+import qualified Data.Text.Encoding.Error as TE
 import Smos.Instance
 import Smos.Web.Server.Foundation
 import Smos.Web.Server.Static
@@ -32,8 +36,16 @@ makeTuiWidget websocketRoute = do
 
 communicateWithSmosInstance :: SmosInstanceHandle -> WebSocketsT Handler ()
 communicateWithSmosInstance sih = do
-  let inputConduit = sourceWS .| debugConduit "Input" .| resizeConduit sih .| smosInstanceInputSink sih
-      outputConduit = smosInstanceOutputSource sih .| debugConduit "Output" .| sinkWSText
+  let inputConduit =
+        sourceWS
+          .| debugConduit "Input"
+          .| resizeConduit sih
+          .| smosInstanceInputSink sih
+      outputConduit =
+        smosInstanceOutputSource sih
+          .| debugConduit "Output"
+          .| sanitiseUtf8Conduit
+          .| sinkWSText
   runConduit (yield ("\ESC[?25h" :: Text) .| sinkWSText) -- turn on cursor
   res <- race (wait (smosInstanceHandleAsync sih)) (race (runConduit inputConduit) (runConduit outputConduit))
   case res of
@@ -57,10 +69,18 @@ resizeConduit instanceHandle = awaitForever $ \bs -> do
     Nothing -> yield bs -- It's not a resize request, just let it through.
     Just terminalSize -> liftIO $ smosInstanceResize instanceHandle terminalSize
 
-debugConduit :: MonadIO m => String -> ConduitT ByteString ByteString m ()
-debugConduit _ = iterMC $ \_ -> do
-  -- liftIO $ threadDelay $ 150 * 1000 -- Artificial delay
+debugConduit :: MonadHandler m => String -> ConduitT ByteString ByteString m ()
+debugConduit name = iterMC $ \bs -> do
+  $(logDebug) (T.pack (name <> ": " <> show bs))
   pure ()
--- debugConduit name = iterMC go
---   where
---     go bs = liftIO $ putStrLn $ name <> ": " <> show bs
+
+-- Just to see what happens if smos sends some non-utf8 for some reason
+-- troubleConduit :: MonadHandler m => ConduitT ByteString ByteString m ()
+-- troubleConduit = awaitForever $ \bs -> do
+--   b <- liftIO $ (> (9 :: Int)) <$> randomRIO (1, 10)
+--   when b $ yield $ SB.pack [0xc0]
+--   yield bs
+
+-- This makes sure that no non-utf8 gets to the client.
+sanitiseUtf8Conduit :: Monad m => ConduitT ByteString Text m ()
+sanitiseUtf8Conduit = C.map (TE.decodeUtf8With TE.ignore)